requirements.txt
telethon
fastapi
uvicorn
supabase
python-dotenv
aiohttp
Pillow
jinja2
python-multipart
requests
